Implemented some options for music selection
So far, the first option is still the "how to play" music as a placeholder, but the "No BGM" option is now functional as well. Options two and three will function the same as the latter.

I also implemented a couple of wrapper functions for the pausing and playing of hUGE music, with a validator to ensure a vblank interrupt doesn't get added or removed multiple times unnecessarily.