Minor fixes, defining next steps.
Next I'm going to be working on implementing Time Attack, but as I write in the readme, I believe the best approach is to modularize some of the existing code in the puzzle-solve scene. It was all in one place for the sake of being a proof-of-concept from my college project, but now that that's old news, it's time to make this into an actual product. Also messing around with some music stuff lol